requirements.txt
# Core LangGraph and LangChain dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.7.0

# Document processing
pypdf2>=3.0.0
pillow>=10.0.0
python-docx>=0.8.11
markdown>=3.5.0

# Code validation and testing
# ast is part of Python standard library
pylint>=3.0.0
flake8>=6.0.0
mypy>=1.7.0
pytest>=7.4.0

# Git integration
gitpython>=3.1.40
PyGithub>=1.59.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
requests>=2.31.0
rich>=13.0.0
typer>=0.9.0

# Code Generation and File Management
pathlib2>=2.3.7
jinja2>=3.1.0
black>=23.0.0
isort>=5.12.0

# MCP and Context Management
mcp>=1.0.0
tiktoken>=0.5.0
networkx>=3.0.0
